Apps/python/example-2/backend/requirements.txt
psycopg2-binary==2.9.9
Flask==2.3.3
flask-cors==4.0.0